Added module service. has database dependencies and application resources
