--- ModulePage/src/main/java/ie/atu/modulepage/Course.java
@@ -1,55 +1,0 @@
-package ie.atu.modulepage;
-
-import jakarta.persistence.*;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Entity
-@Table(name = "course")
-@Data
-@NoArgsConstructor
-
-public class Course {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, unique = true)
-    private String name;
-
-    @ElementCollection
-    private List<Long> moduleIds; // List of Module IDs related to this course
-
-    public Course(Long id, String name, List<Long> moduleIds) {
-        this.id = id;
-        this.name = name;
-        this.moduleIds = moduleIds;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public List<Long> getModuleIds() {
-        return moduleIds;
-    }
-
-    public void setModuleIds(List<Long> moduleIds) {
-        this.moduleIds = moduleIds;
-    }
-}
