+++ ModulePage/src/main/java/ie/atu/modulepage/Module.java
@@ -2,10 +2,8 @@ package ie.atu.modulepage;
 
 import jakarta.persistence.*;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 
 @Entity
-@NoArgsConstructor
 @Data
 public class Module{
 
@@ -16,13 +14,50 @@ public class Module{
     @Column(nullable = false)
     private String name;
 
+    private String description;
+
     @Column(nullable = false)
     private Long courseId; // Links to Course ID
 
-    public Module(Long id, String name, Long courseId) {
+    public Module(){
+
+    }
+    public Module(Long id, String name, Long courseId, String description) {
         this.id = id;
         this.name = name;
         this.courseId = courseId;
+        this.description=description;
     }
 
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public Long getCourseId() {
+        return courseId;
+    }
+
+    public void setCourseId(Long courseId) {
+        this.courseId = courseId;
+    }
 }
