+++ ModulePage/src/main/java/ie/atu/modulepage/ModulePageApplication.java
@@ -3,6 +3,7 @@ package ie.atu.modulepage;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
+
 @SpringBootApplication
 public class ModulePageApplication {
