+++ ModulePage/src/main/resources/application.properties
@@ -1,2 +1,13 @@
 spring.application.name=ModulePage
 server.port=8081
+
+# MySQL Database Configuration
+spring.datasource.url=jdbc:mysql://localhost:3306/module_db
+spring.datasource.username=root
+spring.datasource.password=password
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+
+# JPA Configuration
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL8Dialect
